Standardize loot modifier classes
